Move everything from testing_utils to test_pipeline_mixin
